#1509 renamed function names + new folder 'forms'
