+++ base/forms/my_message.py
@@ -1,0 +1,40 @@
+##############################################################################
+#
+#    OSIS stands for Open Student Information System. It's an application
+#    designed to manage the core business of higher education institutions,
+#    such as universities, faculties, institutes and professional schools.
+#    The core business involves the administration of students, teachers,
+#    courses, programs and so on.
+#
+#    Copyright (C) 2015-2016 Université catholique de Louvain (http://www.uclouvain.be)
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU General Public License as published by
+#    the Free Software Foundation, either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU General Public License for more details.
+#
+#    A copy of this license - GNU General Public License - is available
+#    at the root of the source code of this program.  If not,
+#    see http://www.gnu.org/licenses/.
+#
+##############################################################################
+from django import forms
+from django.utils.translation import ugettext_lazy as _
+
+
+class MyMessageForm(forms.Form):
+    selected = forms.BooleanField(initial=False, required=False)
+    subject = forms.CharField(required=False)
+    created = forms.DateTimeField(required=False)
+    read = forms.BooleanField(initial=False, required=False)
+    id = forms.IntegerField()
+
+
+class MyMessageActionForm(forms.Form):
+    ACTIONS = (('MARK_AS_READ', _('mark_selected_as_read')), ('DELETE', _('delete_selected')))
+    action = forms.ChoiceField(required=False, choices=ACTIONS, initial=None)
