+++ base/models/offer_year_calendar.py
@@ -63,7 +63,7 @@ class OfferYearCalendar(models.Model):
         return u"%s - %s" % (self.academic_calendar, self.offer_year)
 
 
-def _create_offer_year_calendar(academic_calendar, offer_year):
+def _create(academic_calendar, offer_year):
     offer_yr_calendar = OfferYearCalendar(academic_calendar=academic_calendar,
                                           offer_year=offer_year,
                                           start_date=academic_calendar.start_date,
@@ -71,65 +71,23 @@ def _create_offer_year_calendar(academic_calendar, offer_year):
     offer_yr_calendar.save()
 
 
-def _create_offer_year_calendars(academic_calendar):
+def _create_from_academic_calendar(academic_calendar):
     academic_yr = academic_calendar.academic_year
     offer_years = offer_year.find_by_academic_year(academic_yr.id)
     for offer_yr in offer_years:
-        _create_offer_year_calendar(academic_calendar, offer_yr)
-
-#
-# def update_offer_year_calendars(academic_calendar):
-#     offer_year_calendars = find_by_academic_calendar(academic_calendar)
-#     for offer_year_calendar in offer_year_calendars:
-#         offer_year_calendar.update_dates(academic_calendar.start_date, academic_calendar.end_date)
+        _create(academic_calendar, offer_yr)
 
 
-def save_offer_year_calendars(academic_calendar):
+def save_from_academic_calendar(academic_calendar):
     offer_year_calendars = find_by_academic_calendar(academic_calendar)
     if offer_year_calendars:
         for offer_year_calendar in offer_year_calendars:
             offer_year_calendar.update_dates(academic_calendar.start_date, academic_calendar.end_date)
     else:
-        _create_offer_year_calendars(academic_calendar)
+        _create_from_academic_calendar(academic_calendar)
 
 
-# def save(academic_cal):
-#     """
-#     It creates an event in the academic calendar of each annual offer when an
-#     event is created in the academic calendar.
-#     """
-#     academic_yr = academic_cal.academic_year
-#     offer_year_list = offer_year.find_by_academic_year(academic_yr.id)
-#     for offer_yr in offer_year_list:
-#         offer_yr_calendar = OfferYearCalendar()
-#         offer_yr_calendar.academic_calendar = academic_cal
-#         offer_yr_calendar.offer_year = offer_yr
-#         offer_yr_calendar.start_date = academic_cal.start_date
-#         offer_yr_calendar.end_date = academic_cal.end_date
-#         offer_yr_calendar.save()
-#
-#
-# def update(academic_cal):
-#     sent_message_error = None
-#     offer_year_calendar_list = find_by_academic_calendar(academic_cal)
-#     if offer_year_calendar_list:
-#         for offer_year_calendar in offer_year_calendar_list:
-#             if offer_year_calendar.customized: # case offerYearCalendar is already customized
-#                 # We update the new start date
-#                 # WARNING : this is TEMPORARY ; a solution for the sync from EPC to OSIS
-#                 #           because the start_date for scores_encodings doesn't exist in EPC
-#                 offer_year_calendar.start_date = academic_cal.start_date
-#                 offer_year_calendar.save()
-#             else:
-#                 offer_year_calendar.start_date = academic_cal.start_date
-#                 offer_year_calendar.end_date = academic_cal.end_date
-#                 offer_year_calendar.save()
-#     else:
-#         save(academic_cal)
-#     return sent_message_error
-
-
-def offer_year_calendar_by_current_session_exam():
+def find_by_current_session_exam():
     return OfferYearCalendar.objects.filter(start_date__lte=timezone.now()) \
                                     .filter(end_date__gte=timezone.now()).first()
 
@@ -138,7 +96,7 @@ def find_by_academic_calendar(academic_cal):
     return OfferYearCalendar.objects.filter(academic_calendar=int(academic_cal.id))
 
 
-def find_offer_year_calendar(offer_yr):
+def find_by_offer_year(offer_yr):
     return OfferYearCalendar.objects.filter(offer_year=offer_yr,
                                             start_date__isnull=False,
                                             end_date__isnull=False).order_by('start_date',
@@ -164,7 +122,7 @@ def find_deliberation_date(session_exam):
     return scores_encodings_end_date
 
 
-def get_max_end_date(academic_calendar_id):
+def find_latest_end_date_by_academic_calendar(academic_calendar_id):
     try:
         return OfferYearCalendar.objects.filter(academic_calendar_id=academic_calendar_id) \
                                         .filter(customized=True) \
